@@ -11,32 +11,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.1",
-    "@angular/cdk": "^6.4.2",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.2",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^7.1.3",
+    "@angular/cdk": "^7.1.1",
+    "@angular/common": "^7.1.3",
+    "@angular/compiler": "^7.1.3",
+    "@angular/core": "^7.1.3",
+    "@angular/forms": "^7.1.3",
+    "@angular/http": "^7.1.3",
+    "@angular/material": "^7.1.1",
+    "@angular/platform-browser": "^7.1.3",
+    "@angular/platform-browser-dynamic": "^7.1.3",
+    "@angular/router": "^7.1.3",
     "@ng-idle/core": "^6.0.0-beta.3",
     "@ng-idle/keepalive": "^6.0.0-beta.3",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/store": "^6.1.0",
-    "angular2-moment": "^1.9.0",
     "core-js": "^2.5.4",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.1",
+    "tslib": "^1.9.0",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.2",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "~7.1.3",
+    "@angular/compiler-cli": "^7.1.3",
+    "@angular/language-service": "^7.1.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -51,6 +49,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.1.6"
   }
 }
